--- .babelrc
@@ -1,3 +1,0 @@
-{
-"presets": ["react-native"]
-}
