--- android/settings.gradle
@@ -1,3 +1,0 @@
-rootProject.name = 'sweeper'
-
-include ':app'
