+++ example/ios/GestureView/AppDelegate.h
