+++ example/ios/GestureView/Images.xcassets/AppIcon.appiconset/Contents.json
