+++ example/ios/GestureView/Info.plist
