+++ example/ios/GestureViewTests/Info.plist
